requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
django-cors-headers>=4.3.1
flake8>=6.1.0
black>=23.12.1
isort>=5.13.2
Pillow>=10.1.0
djangorestframework
gunicorn
psycopg2-binary
djangorestframework-simplejwt
python-decouple
requests
django-otp==1.0.0
sendgrid-django
djangorestframework
